streamlined implementation of chicanery
I have fixed up menu support, squashed a particularly bad memory access
violation bug, and migrated the initialization of the draw routines to
chicanery.c. I realize that the initialization of these subsystems should not
be the jobs of the expert system but the program that starts up the expert
system. These options may be migrated back at some point but currently this is
the best way to do this.